requirements_mryan.txt
# Add CUDA-enabled PyTorch index while keeping default PyPI index
--extra-index-url https://download.pytorch.org/whl/cu118

# Base requirements for GUI (CPU)
rp
easydict
oldest-supported-numpy
opencv-contrib-python
imageio-ffmpeg
fire
moviepy
icecream
matplotlib
art

# Additional GPU requirements
GitPython
einops
diffusers
transformers
accelerate
peft
sentencepiece
ffmpeg-python

# PyTorch packages
torch
torchvision

# Additional dependencies for Gradio UI
gradio==4.44.1
Pillow
